Fix occasionally appending junk to packets
In response to a Kad search if we had more than 50 results to send, all but the first
package could have junk appended at the end. Though it didn't make any harm on the
receiving client, using bandwidth to send/receive junk is not right.

Thanks for the sharp eyes of Stu Redman.

:: upstream revision 10443 ::