Rewritten CFormat class to support indexed argument references
Excerpt from the class description:

CFormat lacks the following capabilities:
 * The "*" width-modifier, because only one argument is fed at a time.
   (Using it will render the format-specifier invalid.)
 * The "n" type, just unsafe and won't be implemented.
 * The "C" and "S" types, which are considered obsolete.
 * The Long Double type, which is extremly slow and shouldn't be used.

Support for the C99 a, A conversions and the non-standard ', I flags
depend on the underlying C library. Do not use them.

Supports the glibc-specific %m conversion on all platforms, where there's
a way to get the error description. If the underlying C library has a
thread-safe way to get the error description, then this conversion is
thread-safe, too.
  (One current exception to this is Windows, since at the moment it is
  only used in the non-Windows parts of amuled.)

Deviations from printf(3):

CFormat tries hard to format the passed POD-type according to the conversion
type. Basic type conversions may take place to accomplish this goal. This
results in formats accepting a variety of types, namely:
 * c, i, d, u, o, x, X accept wxChar and all integer types,
 * a, A, e, E, f, F, g, G accept wxChar, integer and floating-point types,
 * p accepts only pointers,
 * s accepts all the above mentioned types in addition to wxString and
   wxChar* types.

The only exception from this rule is integer (d, i, u) conversion. It will
always use the correct conversion (i or u) depending on the signedness of
the passed argument.

's' conversions are inspired by "we're converting to string, anyway". Thus
they use a 'default' conversion for each accepted type: 'c' for wxChar, 'i'
and 'u' for signed and unsigned integers, respectively, 'g' for
floating-point numbers and 'p' for pointers.
  (This means that '%s' alone fits for almost all needs, but doing so in
  translated strings would confuse translators. Please use correct
  conversions for better readability.)

Other relaxations / differences from printf(3):
 * Length modifiers are read and validated, but always ignored.
 * As a consequence, invalid combinations of length modifiers and conversion
   types are silently ignored (i.e. for example the invalid '%qs'
   format-specifier is silently treated as '%s').
 * 'p' conversion ignores all modifiers except the argument index reference.
 * You can mix positional and indexed argument references.
 * With indexed argument references we allow to leave gaps in the indices.

:: upstream revision 10459 ::