Use the C++ compiler for feature tests
Most of the configure test results will be used in C++ sources, so testing for
them with the C compiler might give wrong results.

I had to fix a few standard tests to work with the C++ compiler. There are
still a small number of tests that require the C compiler and should not be
checked with the C++ compiler. They either check features of the C compiler
or their results will only be used in C source files (by the C compiler).

This should fix issue #1572

:: upstream revision 10838 ::